\name{readbinarydosage-package}
\alias{readbinarydosage-package}
\alias{readbinarydosage}
\docType{package}
\title{
\packageTitle{readbinarydosage}
}
\description{
\packageDescription{readbinarydosage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{readbinarydosage}
\packageIndices{readbinarydosage}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{readbinarydosage}

Maintainer: \packageMaintainer{readbinarydosage}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
