% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readbinarydosage.R
\name{readsnps}
\alias{readsnps}
\title{readsnps}
\usage{
readsnps(bdfile)
}
\arguments{
\item{bdfile}{Name of binary dosage data file}
}
\value{
Returns status of reading file
}
\description{
Test is the file can be opened
}
\examples{
testfile <- system.file("extdata",
                        "test1.bdose",
                        package = "readbinarydosage")
readsnps(bdfile = testfile)
}
